feat: Implement News API integration with Retrofit
This commit introduces the `NewsApi` interface for interacting with the News API. It defines the `getEvrything` function for fetching news articles, using Retrofit to make a GET request to the `v2/everything` endpoint.
A corresponding `ApiModel` is created to handle data classes (Root, Article, Source) to be received from API.
`getEvrything` function now accept parameters query, and apikey.