feat: Implement dynamic news fetching with query parameter
This commit updates the news fetching functionality to allow for dynamic queries.
`GetEverythingUseCase` now accepts a `query` parameter.
`NewsApiRepositoryImpl` now passes the `query` parameter to the `getEverything` function of the `NewsApi`.
`HomeScreenViewModel` now includes a `getEverything` function that can receive a `query` parameter and updates the UI accordingly.
`HomeScreen` composable has been updated with a `TextField` to set the query, and changes news based on new value of the query.
`getEverything` function is now called on init of `HomeScreenViewModel`.